@@ -8,7 +8,7 @@ const _cyan = '\u001b[36m'
 const _white = '\u001b[37m'
 const reset = '\u001b[0m'
 
-module.exports = {
+export default {
   success: green,
   warning: yellow,
   info: magenta,
