@@ -1,12 +1,17 @@
-const path = require('path')
-const fs = require('fs')
-const _ = require('lodash')
-const prompt = require('prompt-sync')()
-const inquirer = require('inquirer')
-const csvParse = require('csv-parse/lib/sync')
-const yaml = require('js-yaml')
+import path from 'path'
+import fs from 'fs'
+import _ from 'lodash'
+import promptSync from 'prompt-sync'
+import inquirer from 'inquirer'
+import yaml from 'js-yaml'
+import url from 'url'
+import { parse as csvParse } from 'csv-parse/sync'
 
-const colors = require('./colors')
+import colors from './colors.js'
+
+const prompt = promptSync()
+
+const __dirname = url.fileURLToPath(new URL('.', import.meta.url))
 
 const getMaxLength = (types, headName) => {
   return _(types)
@@ -149,7 +154,7 @@ const generateDefinitionsFile = (definitionsFileName, answers, dryRun = false) =
     },
   }
 
-  writeFile(path.join('.', definitionsFileName), yaml.safeDump(content), dryRun)
+  writeFile(path.join('.', definitionsFileName), yaml.dump(content), dryRun)
 }
 
 const promptBool = (text, defaultValue = true) => {
@@ -198,10 +203,12 @@ const writeFile = (path, text, dryRun = false) => {
 //
 // Main
 //
-module.exports = async (program, templateFileName, definitionsFileName) => {
+export default async (program, templateFileName, definitionsFileName) => {
+  const options = program.opts()
+
   await inquirer.prompt(createQuestions()).then((answers) => {
-    generateTemplateFile(templateFileName, answers, program.dryRun)
-    generateDefinitionsFile(definitionsFileName, answers, program.dryRun)
+    generateTemplateFile(templateFileName, answers, options.dryRun)
+    generateDefinitionsFile(definitionsFileName, answers, options.dryRun)
 
     console.log()
     console.log(`${colors.success}Generated '${templateFileName}' and '${definitionsFileName}'.${colors.reset}`)
