@@ -1,20 +1,21 @@
 #!/usr/bin/env node
 
-const program = require('commander')
-const path = require('path')
-const fs = require('fs')
-const _ = require('lodash')
-const execSync = require('child_process').execSync
-const yaml = require('js-yaml')
+import path from 'path'
+import fs from 'fs'
+import _ from 'lodash'
+import yaml from 'js-yaml'
+import { Command } from 'commander'
+import { execSync } from 'child_process'
 
-const colors = require('./colors')
-const main = require('./main')
-const genConfig = require('./gen-config')
-const pk = require('../package.json')
+import colors from './colors.js'
+import main from './main.js'
+import genConfig from './gen-config.js'
 
 const templateFileName = '.gitcommit_template'
 const definitionsFileName = '.git_consistent'
 
+const program = new Command()
+const pk = JSON.parse(fs.readFileSync('./package.json'))
 const version = pk.version
 
 //
@@ -35,18 +36,16 @@ const getFilePath = (filePaths, fileName) => {
 }
 
 const loadYaml = (path) => {
-  return yaml.safeLoad(fs.readFileSync(path, 'utf8'))
+  return yaml.load(fs.readFileSync(path, 'utf8'))
 }
 
 const setCustomOptions = (program, definitions) => {
-  program.option('', '---- custom options ----')
   _.forEach(definitions, (definition, term) => {
     const optionName = term == 'subject' ? `-m, --${term}` : `--${term}`
     const valueStr = definition.required ? `<${term}>` : `[${term}]`
     const defaultValue = definition.default
     if (!['variable', 'branch'].includes(definition.type)) program.option(`${optionName} ${valueStr}`, definition.description, defaultValue)
   })
-  program.option('', '---- custom options ----')
 }
 
 const showErrorMessageAndExit = (e) => {
@@ -92,6 +91,7 @@ program
   .option('-D, --dry-run', 'run dry-run mode')
   .option('-i, --interactive', 'run interactive mode')
   .option('-S, --silent', "don't show commit command")
+  .option('-I, --init', 'generate config files')
   .on('--help', () => {
     console.log('')
     console.log('File Paths:')
@@ -100,15 +100,7 @@ program
   })
   .version(version)
 
-program
-  .command('init')
-  .description('generate config files')
-  .option('-D, --dry-run', 'run dry-run mode')
-  .action(() => {
-    /* no-op */
-  })
-
-if (process.argv[2] === 'init') {
+if (['-I', '--init'].includes(process.argv[2])) {
   runInit(program)
 } else if (['-h', '--help'].includes(process.argv[2])) {
   runHelp(program, definitionsFilePath)
