@@ -1,9 +1,9 @@
-const _ = require('lodash')
-const execSync = require('child_process').execSync
-const inquirer = require('inquirer')
-const emoji = require('node-emoji')
+import _ from 'lodash'
+import inquirer from 'inquirer'
+import emoji from 'node-emoji'
+import { execSync } from 'child_process'
 
-const colors = require('./colors')
+import colors from './colors.js'
 
 //
 // Functions
@@ -171,29 +171,29 @@ const createQuestion = (definition, term) => {
   }
 }
 
-const createQuestions = (program, definitions) => {
+const createQuestions = (options, definitions) => {
   return _(definitions)
     .map((definition, term) => {
-      if (!_.isUndefined(program[term]) || !_.isUndefined(definition.default)) return
+      if (!_.isUndefined(options[term]) || !_.isUndefined(definition.default)) return
       return createQuestion(definitions[term], term)
     })
     .compact()
     .value()
 }
 
-const getValue = (program, definition, answer, term) => {
+const getValue = (options, definition, answer, term) => {
   const val = branchText(definition)
 
   switch (definition.type) {
     case 'enum':
     case 'string':
     case 'text':
-      if (!_.isUndefined(program[term])) return program[term]
+      if (!_.isUndefined(options[term])) return options[term]
       if (!_.isUndefined(answer)) return answer
       if (!_.isUndefined(definition.default)) return definition.default
       return ''
     case 'variable':
-      if (!_.isUndefined(program[definition.origin])) return program[definition.origin]
+      if (!_.isUndefined(options[definition.origin])) return options[definition.origin]
       return ''
     case 'branch':
       if (!_.isUndefined(val)) return val
@@ -203,17 +203,16 @@ const getValue = (program, definition, answer, term) => {
   }
 }
 
-const replaceTerms = async (program, template, definitions, callback) => {
+const replaceTerms = async (options, template, definitions, callback) => {
   await inquirer
-    .prompt(createQuestions(program, definitions))
+    .prompt(createQuestions(options, definitions))
     .then((answers) => {
       let result = template
       _.forEach(definitions, (definition, term) => {
         const answer = answers[term]
-        const value = getValue(program, definition, answer, term).trim()
+        const value = getValue(options, definition, answer, term).trim()
 
-        program[term] = value
-
+        options[term] = value
         const errorMessages = checkValue(term, value, definition)
         if (!_.isEmpty(errorMessages)) throw new Error(errorMessages)
 
@@ -256,10 +255,12 @@ const gitCommit = (commitMessage, duet = false, silent = false, dryRun = false) 
 //
 // Main
 //
-module.exports = async (program, template, definitions) => {
-  if (!program.dryRun) checkAddedFile()
+export default async (program, template, definitions) => {
+  const options = program.opts()
 
-  await replaceTerms(program, template, definitions, (commitMessage) => {
-    gitCommit(commitMessage.trim(), program.duet, program.silent, program.dryRun)
+  if (!options.dryRun) checkAddedFile()
+
+  await replaceTerms(options, template, definitions, (commitMessage) => {
+    gitCommit(commitMessage.trim(), options.duet, options.silent, options.dryRun)
   })
 }
